==> 01 Preprocess and model.Rmd <==
---
title: "01 Preprocess and model"
author: "Dominik Klepl"
date: "4/22/2020"
output: html_document
---

```{r}
library(eegkit)
library(ggplot2)
library(ggthemes)
library(tidyr)
library(dplyr)
```

```{r}
files = list.files("data/merged", full.names = T)
```

## Try on sample + visualize
```{r}
df = read.csv(files[8])
```

Transform to mV
```{r}
df[,1:32] = lapply(df[,1:32], function(x) 1e-6*x)
```

Plot EEG cap
```{r}
#eegcap(electrodes = colnames(df)[-1], type="3d")
png("figures/eeg_cap.png",res=125)
eegcap(electrodes = colnames(df)[1:32], type="2d")
dev.off()
```

## Filter
```{r}
sampling = 500
filtered = df
filtered[,1:32]= eegfilter(df[,1:32], sampling, lower = 7, upper = 30, method = "butter", order = 5, forwardreverse = F)
```

Visualize effect of filter
```{r}
filter_effect = data.frame(Raw = df$Cz,
                           Filter = filtered$Cz,
                           time = 1:nrow(df))
filter_effect = filter_effect[1:(sampling*60),]
filter_effect = tidyr::gather(filter_effect,Signal,mV,Raw:Filter)

(f_e_plot = ggplot(filter_effect,aes(time,mV,color=Signal))+
  geom_line()+
  facet_wrap(~Signal)+
  theme_few()+
  scale_color_tableau())
ggsave("figures/Raw_filtered.png",f_e_plot,width=7,height=4)
rm(f_e_plot)
rm(filter_effect)
```

### Normalize
```{r}
filtered[,1:32] = lapply(filtered[,1:32], function(x) scale(x))

#sanity check - mean should be 0 and SD 1
mean(filtered$Fp1)
sd(filtered$Fp1)

#resample to 200 Hz
filtered = as.data.frame(eegresample(filtered,round(nrow(filtered)/2.5,0)))
filtered[,33:39] = lapply(filtered[,33:39], function(x) ifelse(x>0.5,1,0))
filtered$Series = round(filtered$Series,0)
```

## Check distribution
```{r}
t = filtered %>% gather()

ggplot(t, aes(value)) + 
  geom_density() + 
  facet_wrap(~key, scales = 'free_x')+
  theme_few()
```

### Linear models

Fit a model for each class
```{r}
EEG_cols = c(1:32,40)
lab_cols = 33:38
mean_auc = rep(NA,length(lab_cols))
for (i in 1:length(lab_cols)) {
  df_rest = filtered[,c(EEG_cols,lab_cols[i])]
  colnames(df_rest)[34]="Response"
  df_rest$Response = as.factor(df_rest$Response)
  
  df_rest = caret::downSample(x=df_rest[,1:33],y=df_rest$Response, yname = "Response")
  
  train = df_rest[df_rest$Series<=6,]
  test = df_rest[df_rest$Series>6,]
  
  train = train[,-33]
  test = test[,-33]
  
  m1 = glm(Response ~ ., family = "binomial", data=train)
  summary(m1)
  
  preds = predict(m1, newdata = test, type = "response")
  #preds = ifelse(preds>0.5,1,0)
  auc = MLmetrics::AUC(preds,test$Response)
  mean_auc[i] = auc
  cat("\n",colnames(df)[lab_cols[i]],":",auc)
}

mean(mean_auc)
```
A baseline model gives 0.66 AUC.

```{r}
cor = round(cor(df[,1:32]),2)

cor.mtest = function(mat, ...) {
    mat = as.matrix(mat)
    n = ncol(mat)
    p.mat = matrix(NA, n, n)
    diag(p.mat) = 0
    for (i in 1:(n - 1)) {
        for (j in (i + 1):n) {
            tmp = cor.test(mat[, i], mat[, j], ...)
            p.mat[i, j] = p.mat[j, i] <- tmp$p.value
        }
    }
  colnames(p.mat) = rownames(p.mat) = colnames(mat)
  p.mat
}

# matrix of the p-value of the correlation
p_vals = cor.mtest(df[,1:32])

corrplot::corrplot(cor,"color",tl.col="black", tl.cex = 0.75, p.mat = p_vals,sig.level = 0.001,addCoef.col = "black",number.cex = 0.5)
```

Put all preprocessing into a function
```{r}
preprocess = function(data, sampling = 500, resample = F, new_sampling = 200) {
  data[,1:32] = lapply(data[,1:32], function(x) 1e-6*x)
  data[,1:32]= eegfilter(data[,1:32], sampling, lower = 7, upper = 30, method = "butter", order = 5, forwardreverse = F)
  if (resample){
    down = sampling/new_sampling
    data = as.data.frame(eegresample(data,round(nrow(data)/down,0)))
    data[,33:39] = lapply(data[,33:39], function(x) ifelse(x>0.5,1,0))
    data$Series = round(data$Series,0)
  }
  data[,1:32] = lapply(data[,1:32], function(x) scale(x))
  return(data)
}
```

## Run Linear models for all subjects

```{r}
makeFolds = function(data) {
  folds = list()
  for (i in unique(data$Series)){
    f = rownames(data[data$Series==i,])
    f = as.integer(f)
    name = paste0("Fold",i)
    folds[[name]] = f
  }
  return(folds)
}
```

```{r}
library(caret)
library(dplyr)
library(eegkit)
files = list.files("data/merged", full.names = T)
files = files[c(2,5,11)]

EEG_cols = c(1:32,40)
lab_cols = 33:38
n=1
scores = list()
for (f in files[-1]) {
  df = read.csv(f)
  df = preprocess(df)
  
  mean = rep(NA,length(lab_cols))
  sd = rep(NA,length(lab_cols))
  for (i in 1:length(lab_cols)) {
    df_rest = df[,c(EEG_cols,lab_cols[i])]
    colnames(df_rest)[34]="Response"
    df_rest$Response = as.factor(df_rest$Response)
    set.seed(1710)
    df_rest = caret::downSample(x=df_rest[,1:33],y=df_rest$Response, yname = "Response")
    levels(df_rest$Response) = c("Null","Response")
    #folds = makeFolds(df_rest)
    
    ctrl = trainControl(method="cv",
                        index = groupKFold(df_rest$Series, 4),
                        classProbs = T, savePredictions = "final")
    m = train(Response ~., data = df_rest,
              method = "glm", trControl = ctrl)
    
    pr = m$pred
    pr$pred = ifelse(pr$pred=="Null",0,1)
    pr$obs = ifelse(pr$obs=="Null",0,1)
    pr = pr[,c(2,3,7)]
    pr = pr %>% group_by(Resample) %>% summarise(F1 = MLmetrics::F1_Score(obs, pred))

    mean[i] = mean(pr$F1)
    sd[i] = sd(pr$F1)
    cat("\n",colnames(df)[lab_cols[i]],":",mean[i])
  }
  
  res = data.frame(mean = mean, sd = sd)
  scores[[strsplit(f,"/")[[1]][3]]] = res
}

save(scores,file = "results/Baseline_results.RData")
```
